+++ src/airConditionerServices.ts
@@ -1,8 +1,8 @@
 import { AirConditioner } from './airConditioner/airConditioner.js';
 import { PlatformAccessory }  from 'homebridge';
 import { SmartHqPlatform }    from './platform.js';
-import { DeviceService }      from 'ge-smarthq';
+import { DeviceService }      from 'ge-smarthq-api';
 
-export function setupAirConditionerServices(this: SmartHqPlatform, accessory: PlatformAccessory, deviceServices: DeviceService[], deviceId: string) {
-  new AirConditioner(this, accessory, deviceServices, deviceId);
+export function setupAirConditionerServices(this: SmartHqPlatform, accessory: PlatformAccessory, deviceServices: DeviceService[], deviceId: string, groupAccessory?: PlatformAccessory[]) {
+  new AirConditioner(this, accessory, deviceServices, deviceId, groupAccessory ?? []);
 }